pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.26.0"
pytest = "^6.2.5"
gunicorn = "^20.1.0"
pymongo = "^4.2.0"
mongomock = "^4.1.2"
Flask = "^2.2.2"
python-dotenv = "^0.20.0"
bson = "^0.5.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
